requirements.txt
geopandas==1.1.2
anthropic==0.94.0
streamlit==1.54.0
folium==0.15.0
streamlit-folium==0.12.0
plotly==6.7.0
PuLP==2.8.0
pyarrow==17.0.0
python-dotenv==1.0.1
requests==2.33.0
shapely==2.0.4
fiona==1.10.1
pyproj==3.6.1
pytest==7.4.4
pytest-mock==3.12.0
pillow==12.1.1
